refactor(vitest-plugin): move instrumented behavior in dedicated file
